Adds Runner that takes list of votes on polls and return metric for each
This is useful when you're loading a CSV with, say:

poll1,poll2,poll3
0,1,1
1,0,1
1,,1

And you want to calculate the Rice Index on each poll. Runner.run() will take
the list of dicts created from that CSV, as:

[
    {'poll1': 0, 'poll2': 1, 'poll3': 1},
    {'poll1': 1, 'poll2': 0, 'poll3': 1},
    {'poll1': 1, 'poll2': None, 'poll3': 1}
]

And run the chosen `metric_method` on each poll individually. If you ran
RiceIndex.calculate(), it would return:

{
    'poll1': 0.3333333333333333,
    'poll2': 0,
    'poll3': 1
}