Allow changing YES/NO default values
This will be useful when we start loading the votes from CSV files with
DictReader. Unfortunately, DictReader doesn't parse the CSV columns as numbers
by default, even if they're all numbers. I could've parsed them myself, but it
was easier to allow using different values to represent YES/NO, so I could
simply set YES = '1', and NO = '0', and be done with it.