RiceIndex = None if calculating with a single vote
This is a bit different from the original RiceIndex. If there's a single vote,
the RiceIndex should be 1. But given that we'll often be using these methods on
just two people, it's odd that if one of the two spends a few months without
voting on any polls we'll consider their cohesion to be 1.

I've read a few articles that argumented that it made some sense. They said
that not voting might mean "I'm OK with whatever the majority chooses", but
there're many other reasons for not voting, so I don't think this makes much
sense in our case.

All in all, the least of evils seems to me to be ignoring polls where just one
person votes (i.e. returning None).