Fix bug when filtering by multiple columns
The problem was that we filtered by each column at a time. This means that, at
first, `votes` had (say) 100 rows. Then, after the first filter, we've left
with 70 rows. When we move on to filter again, as we're using the `metadata`
array, which still has 100 rows, we would start filtering out the wrong rows.

The solution in this patch is to build the filter all at once and only apply at
the last step.