Adds groupby functionality to Runner.main()
Say that we have the following data:

name,party,poll1,poll2
Joao,PT,1,1
Pedro,PT,1,1
Luana,PT,1,1
Carlos,PT,1,1
Maria,PSOL,0,0

If we were to calculate the rice index between parties PT and PSOL, we would
get:

poll1,poll2
0.6,0.6

Which feels wrong, given that the PSOL party always voted against the PT party
(in this example). The problem is that there're 4 people on the PT party and
only 1 on PSOL, so the data is very skewed. To correct this, what I've done was
create a `groupby` option, so if you run:

./bin/rice_index my_data.csv --party PT PSOL --groupby party

It would group the data by party before calculating the metrics (but after
filtering it). The idea is getting what the median vote for each party. So we
would have:

party,poll1,poll2
PT,1,1
PSOL,0,0

And we would calculate the metrics over this data, which would give us the
expected values:

poll1,poll2
0,0

When calculating the median vote of a party, we only consider the data from the
people that actually voted (ignoring NULL values). If there's a tie (i.e. 50%
votes YES and 50% votes NO), we take any one (not guaranteed which). Maybe, in
the future, it makes more sense to consider a NULL vote when the party is
perfectly divided.