Filter out "S.Part." when genetaring coalitions change list
It creates a bug. For example, Luciana Genro used to be from PT, but was kicked
out and created a new party (PSOL). On PT, she was inside the coalition, on
PSOL she is outside it. The problem is that between the time she was kicked out
PT and before creating PSOL, she voted on a few rollcalls. In these rollcalls,
her party was "S.Part.", which is outside of the coalition (as it's not a party
at all). But as the code doesn't know that, it looks like she was on PT, then
moved to the party "S.Part." that's out of the coalition, then moved to PSOL
that's also out of the coalition, but as there wasn't a coalition change
between her time at "S.Part." and PSOL, it would only print her time at
"S.Part.".

We could have treated this in many ways. For example, we could consider that
when a person is "S.Part.", she actually is on the following party, but haven't
got into filling up the papers yet, so in Luciana's case we should consider she
on PSOL from the day she was kicked out of PT. This might make more sense, but
it'll have to be tested.

Let's keep it simple for now.